requirements.txt
telegram
langchain_community
langchain_openai
langchain
python-telegram-bot
pandas
llama-index
python-dotenv
openai
tiktoken